Redefined model classes and started Controllers

+++ src/main/java/br/com/buscadevapi/repository/SkillRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface SkillRepository extends JpaRepository<Skill, Long> {
-    List<Skill> findAllByIdIn(List<Integer> list);
+    List<Skill> findAllByIdIn(List<Long> list);
 }
